+++ plugins/eta.js
@@ -1,5 +1,0 @@
-export { default } from "./eta.ts";
-const file = "lume/plugins/eta";
-console.log(
-  `WARNING: Deprecated Javascript file ${file}.js. Use the Typescript version ${file}.ts`,
-);