+++ plugins/terser.js
@@ -1,5 +1,0 @@
-export { default } from "./terser.ts";
-const file = "lume/plugins/terser";
-console.log(
-  `WARNING: Deprecated Javascript file ${file}.js. Use the Typescript version ${file}.ts`,
-);